refactor: split verification_constraint file
Move each verifier into its own `.rs` file. The original
`verification_constraint.rs` file was too big.

Note: the public API is not affected by this refactor. Everything keeps
working as expected for the consumers of our library.
